reorganized file structure, implementing badge detail page
