+++ src/App.js
@@ -1,9 +1,10 @@
 import './App.css';
 import './styles/style.scss';
-import ResourceCatalog from "./components/ExistingInfrastructure/ResourceCatalog/ResourceCatalog";
+import ResourceCatalog from "./pages/ResourceCatalog";
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import ResourceDetail from "./components/ExistingInfrastructure/ResourceDetail/ResourceDetail";
-import BadgeDetail from "./components/ExistingInfrastructure/BadgeDetail/BadgeDetail";
+import ResourceDetail from "./pages/ResourceDetail";
+import BadgeDetail from "./pages/BadgeDetail";
+import Demo from "./pages/Demo";
 
 function App() {
     return (
@@ -13,6 +14,10 @@ function App() {
                     <Route path="/" element={<ResourceCatalog />} />
                     <Route path="/resourceDetail/:resourceId" element={<ResourceDetail />} />
                     <Route path="/resourceBadge/:resourceId/:badgeId" element={<BadgeDetail />} />
+
+                    {process.env.NODE_ENV === "production" ? null : (
+                        <Route path="/demo" element={<Demo />} />
+                    )}
                 </Routes>
             </BrowserRouter>
         </div>
