--- src/components/ExistingInfrastructure/BadgeDetail/BadgeDetailSideBar.jsx
@@ -1,7 +1,0 @@
-export default function BadgeDetailSideBar() {
-    return (
-        <div>
-
-        </div>
-    );
-}
