--- src/components/ExistingInfrastructure/ResourceDetail/BadgeSection/BadgeContainer/ResourceBadge/ResourceBadgeRPTag.jsx
@@ -1,39 +1,0 @@
-import {useEffect, useState} from "react";
-
-export default function ResourceBadgeRPTag ({title}) {
-    const [tag, setTag] = useState(title);
-    const [className, setClassName] = useState("badge resource-badge-tag");
-
-    useEffect(() => {
-        let newTag = title;
-        let newClassName = "badge resource-badge-tag";
-
-        // Determine the new tag name and class based on conditions
-        if (title === "Deprecated") {
-            newClassName += " deprecated-style";
-        } else if (title === "NotStarted") {
-            newTag = "Not Started";
-            newClassName += " not-started-style";
-        } else if (title === "Planned") {
-            newTag = "In Progress";
-            newClassName += " planned-style";
-        } else if (title === "Verified") {
-            newTag = "Badge Available";
-            newClassName += " verified-style";
-        } else if (title === "TaskCompleted") {
-            newTag = "Pending Verification";
-            newClassName += " task-completed-style";
-        } else if (title === "VerificationFailed") {
-            newTag = "Verification Failed";
-            newClassName += " verification-failed-style";
-        }
-
-        // Set the state with the new values
-        setTag(newTag);
-        setClassName(newClassName);
-    }, [title]);
-
-    return (
-        <span className={className}>{tag}</span>
-    );
-}
