+++ src/components/badgeDetail/content/BadgeDetailContent.jsx
@@ -1,0 +1,9 @@
+import BadgeDetailBasicInfo from "./BadgeDetailBasicInfo";
+
+export default function BadgeDetailContent({data}) {
+    return (
+        <div className="content-wrapper">
+            <BadgeDetailBasicInfo data={data}/>
+        </div>
+    );
+}
