+++ src/components/resourceCatalog/titleSection/CatalogTitle.jsx
