+++ src/components/resourceDetail/basicInfoSection/BasicInfoFeatures.jsx
